comments and refactor
