+++ packages/rs-dpp/src/state_transition/state_transitions/document/batch_transition/methods/mod.rs
@@ -1068,7 +1068,7 @@ impl DocumentsBatchTransitionMethodsV1 for BatchTransition {
         data_contract_id: Identifier,
         token_contract_position: u16,
         amount: TokenAmount,
-        agreed_total_cost: Credits,
+        total_agreed_price: Credits,
         identity_public_key: &IdentityPublicKey,
         identity_contract_nonce: IdentityNonce,
         user_fee_increase: UserFeeIncrease,
@@ -1100,7 +1100,7 @@ impl DocumentsBatchTransitionMethodsV1 for BatchTransition {
                     data_contract_id,
                     token_contract_position,
                     amount,
-                    agreed_total_cost,
+                    total_agreed_price,
                     identity_public_key,
                     identity_contract_nonce,
                     user_fee_increase,
