+++ packages/rs-dpp/src/state_transition/state_transitions/document/batch_transition/v1/mod.rs
@@ -4,6 +4,7 @@ mod json_conversion;
 mod state_transition_like;
 mod types;
 mod v0_methods;
+mod v1_methods;
 #[cfg(feature = "state-transition-value-conversion")]
 mod value_conversion;
 mod version;
